+++ tests/_internal/modules/test_security.py
@@ -408,7 +408,7 @@ class TestSecurityModule(object):
         mock_storage_client.security = mock_storage_security_client
         response = mocker.MagicMock(spec=Response)
         response.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.return_value = response
+        mock_storage_security_client.get_plan_security_events.return_value = response
         storage_client_factory.get_storage_client_from_plan_uid.return_value = mock_storage_client
         security_module = SecurityModule(
             security_client_one_location,
@@ -418,7 +418,7 @@ class TestSecurityModule(object):
         )
         for page, cursor in security_module.get_user_security_events("foo"):
             pass
-        mock_storage_security_client.get_security_detection_events_for_plan.assert_called_once_with(
+        mock_storage_security_client.get_plan_security_events.assert_called_once_with(
             "111111111111111111",
             cursor=None,
             event_types=None,
@@ -442,7 +442,7 @@ class TestSecurityModule(object):
         response1.text = '{"data": {"cursor": "1:1"}}'
         response2 = mocker.MagicMock(spec=Response)
         response2.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.side_effect = [
+        mock_storage_security_client.get_plan_security_events.side_effect = [
             response1,
             response2,
         ]
@@ -455,7 +455,7 @@ class TestSecurityModule(object):
         )
         for page, cursor in security_module.get_user_security_events("foo"):
             pass
-        assert mock_storage_security_client.get_security_detection_events_for_plan.call_count == 2
+        assert mock_storage_security_client.get_plan_security_events.call_count == 2
 
     def test_get_user_security_events_when_mutliple_plans_returned_calls_security_client_expected_number_of_times(
         self,
@@ -470,7 +470,7 @@ class TestSecurityModule(object):
         mock_storage_client.security = mock_storage_security_client
         response = mocker.MagicMock(spec=Response)
         response.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.return_value = response
+        mock_storage_security_client.get_plan_security_events.return_value = response
         storage_client_factory.get_storage_client_from_plan_uid.return_value = mock_storage_client
         security_module = SecurityModule(
             security_client_two_plans_one_node,
@@ -480,7 +480,7 @@ class TestSecurityModule(object):
         )
         for page, cursor in security_module.get_user_security_events("foo"):
             pass
-        assert mock_storage_security_client.get_security_detection_events_for_plan.call_count == 2
+        assert mock_storage_security_client.get_plan_security_events.call_count == 2
 
     def test_get_user_security_events_when_mutliple_plans_with_cursors_returned_calls_security_client_expected_number_of_times(
         self,
@@ -497,7 +497,7 @@ class TestSecurityModule(object):
         response1.text = '{"data": {"cursor": "1:1"}}'
         response2 = mocker.MagicMock(spec=Response)
         response2.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.side_effect = [
+        mock_storage_security_client.get_plan_security_events.side_effect = [
             response1,
             response2,
             response1,
@@ -512,7 +512,7 @@ class TestSecurityModule(object):
         )
         for page, cursor in security_module.get_user_security_events("foo"):
             pass
-        assert mock_storage_security_client.get_security_detection_events_for_plan.call_count == 4
+        assert mock_storage_security_client.get_plan_security_events.call_count == 4
 
     def test_get_plan_security_events_calls_security_client_with_expected_params(
         self,
@@ -527,7 +527,7 @@ class TestSecurityModule(object):
         mock_storage_client.security = mock_storage_security_client
         response = mocker.MagicMock(spec=Response)
         response.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.return_value = response
+        mock_storage_security_client.get_plan_security_events.return_value = response
         storage_client_factory.get_storage_client_from_plan_uid.return_value = mock_storage_client
         security_module = SecurityModule(
             security_client, storage_client_factory, file_event_client_factory, alert_client_factory
@@ -536,7 +536,7 @@ class TestSecurityModule(object):
             PlanStorageInfo("111111111111111111", "41", "4")
         ):
             pass
-        mock_storage_security_client.get_security_detection_events_for_plan.assert_called_once_with(
+        mock_storage_security_client.get_plan_security_events.assert_called_once_with(
             "111111111111111111",
             cursor=None,
             event_types=None,
@@ -560,7 +560,7 @@ class TestSecurityModule(object):
         response1.text = '{"data": {"cursor": "1:1"}}'
         response2 = mocker.MagicMock(spec=Response)
         response2.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.side_effect = [
+        mock_storage_security_client.get_plan_security_events.side_effect = [
             response1,
             response2,
         ]
@@ -572,7 +572,7 @@ class TestSecurityModule(object):
             PlanStorageInfo("111111111111111111", "41", "4")
         ):
             pass
-        assert mock_storage_security_client.get_security_detection_events_for_plan.call_count == 2
+        assert mock_storage_security_client.get_plan_security_events.call_count == 2
 
     def test_get_plan_security_events_when_mutliple_plans_returned_calls_security_client_expected_number_of_times(
         self,
@@ -587,7 +587,7 @@ class TestSecurityModule(object):
         mock_storage_client.security = mock_storage_security_client
         response = mocker.MagicMock(spec=Response)
         response.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.return_value = response
+        mock_storage_security_client.get_plan_security_events.return_value = response
         storage_client_factory.get_storage_client_from_plan_uid.return_value = mock_storage_client
         security_module = SecurityModule(
             security_client, storage_client_factory, file_event_client_factory, alert_client_factory
@@ -598,7 +598,7 @@ class TestSecurityModule(object):
         ]
         for page, cursor in security_module.get_plan_security_events(plans):
             pass
-        assert mock_storage_security_client.get_security_detection_events_for_plan.call_count == 2
+        assert mock_storage_security_client.get_plan_security_events.call_count == 2
 
     def test_get_plan_security_events_when_mutliple_plans_with_cursors_returned_calls_security_client_expected_number_of_times(
         self,
@@ -615,7 +615,7 @@ class TestSecurityModule(object):
         response1.text = '{"data": {"cursor": "1:1"}}'
         response2 = mocker.MagicMock(spec=Response)
         response2.text = '{"data": {}}'
-        mock_storage_security_client.get_security_detection_events_for_plan.side_effect = [
+        mock_storage_security_client.get_plan_security_events.side_effect = [
             response1,
             response2,
             response1,
@@ -631,7 +631,7 @@ class TestSecurityModule(object):
         ]
         for page, cursor in security_module.get_plan_security_events(plans):
             pass
-        assert mock_storage_security_client.get_security_detection_events_for_plan.call_count == 4
+        assert mock_storage_security_client.get_plan_security_events.call_count == 4
 
     # the order the items are iterated through is not deterministic in some versions of python,
     # so we simply test that the value returned is one of the _possible_ values.